+++ src/Components/MainTableView/UpdateTuple/UpdateTuple.tsx
@@ -15,7 +15,15 @@ type updateTupleState = {
   updateAccessible: boolean // disables submit button if any of the dependencies are inaccessible
 }
 
-class UpdateTuple extends React.Component<{token: string, selectedSchemaName:string, selectedTableName: string, tableAttributesInfo?: TableAttributesInfo, fetchTableContent: any, tupleToUpdate?: any, clearEntrySelection: any}, updateTupleState> {
+class UpdateTuple extends React.Component<{
+    token: string, 
+    selectedSchemaName:string, 
+    selectedTableName: string, 
+    tableAttributesInfo?: TableAttributesInfo, 
+    fetchTableContent: any, 
+    clearEntrySelection: any,
+    selectedTableEntry: any}, 
+  updateTupleState> {
   constructor(props: any) {
     super(props);
     this.state = {
@@ -49,16 +57,10 @@ class UpdateTuple extends React.Component<{token: string, selectedSchemaName:str
       }
     }
 
-    let tupleBuffer = Object.assign({}, this.state.tupleBuffer);
-      Object.values(this.props.tupleToUpdate).forEach((columns: any) => {
-        Object.entries(columns.primaryEntries).forEach((attributeKeyVal: any) => {
-          tupleBuffer[attributeKeyVal[0]] = attributeKeyVal[1]
-        })
-        Object.entries(columns.secondaryEntries).forEach((attributeKeyVal: any) => {
-          tupleBuffer[attributeKeyVal[0]] = attributeKeyVal[1]
-        })
-      })
-      this.setState({tupleBuffer: tupleBuffer});
+    // Check if there is a selectedTableEntry, if so update tuple buffer with it
+    if (this.props.selectedTableEntry !== undefined) {
+      this.setState({tupleBuffer: this.props.selectedTableEntry});
+    }
   }
 
   /**
@@ -68,19 +70,11 @@ class UpdateTuple extends React.Component<{token: string, selectedSchemaName:str
    */
   componentDidUpdate(prevProps: any, prevState: any) {
     // break if there has been no change to the tuple selection 
-    if (this.props.tupleToUpdate === prevProps.tupleToUpdate) {
+    if (this.props.selectedTableEntry === prevProps.selectedTableEntry || this.props.selectedTableEntry === undefined) {
       return;
-    } else {
-      let tupleBuffer = Object.assign({}, this.state.tupleBuffer);
-      Object.values(this.props.tupleToUpdate).forEach((columns: any) => {
-        Object.entries(columns.primaryEntries).forEach((attributeKeyVal: any) => {
-          tupleBuffer[attributeKeyVal[0]] = attributeKeyVal[1]
-        })
-        Object.entries(columns.secondaryEntries).forEach((attributeKeyVal: any) => {
-          tupleBuffer[attributeKeyVal[0]] = attributeKeyVal[1]
-        })
-      })
-      this.setState({tupleBuffer: tupleBuffer});
+    } 
+    else {
+      this.setState({tupleBuffer: this.props.selectedTableEntry});
     }
   }
 
@@ -89,7 +83,7 @@ class UpdateTuple extends React.Component<{token: string, selectedSchemaName:str
    * @param attributeName Attribute name of the change, this is used to access the tupleBuffer object members to set the value
    * @param event Event object that come from the onChange function
    */
-  handleChange(attributeName: string, event: any) {
+  handleChange(event: any, attributeName: string) {
     // Create a copy, update the object, then set state
     let tupleBuffer = Object.assign({}, this.state.tupleBuffer);
     tupleBuffer[attributeName] = event.target.value;
@@ -212,7 +206,7 @@ class UpdateTuple extends React.Component<{token: string, selectedSchemaName:str
     return (
       <div className="updateActionContainer">
         {
-            Object.entries(this.props.tupleToUpdate).length === 0 ?
+            this.props.selectedTableEntry === undefined ?
               <p>Select a table entry from below to update</p>
             :
             <form onSubmit={this.onSubmit}>
@@ -245,7 +239,7 @@ class UpdateTuple extends React.Component<{token: string, selectedSchemaName:str
                              selectedSchemaName={this.props.selectedSchemaName}
                              selectedTableName={this.props.selectedTableName}
                              tableAttributesInfo={this.props.tableAttributesInfo}
-                             tupleToCheckDependency={Object.values(this.props.tupleToUpdate)}
+                             tupleToCheckDependency={Object.values(this.props.selectedTableEntry)}
                              clearList={!Object.entries(this.state.dependencies).length}
                              dependenciesReady={(depList: Array<any>) => this.handleDependencies(depList)} 
                              allAccessible={(bool: boolean) => this.setState({updateAccessible: bool})} />